removed WpApiParserHandler, see #15
